+++ bench/stream.exs
@@ -6,45 +6,45 @@ scheme = System.get_env("CH_SCHEME") || "http"
 
 {:ok, conn} = Ch.start_link(scheme: scheme, hostname: hostname, port: port)
 
-statement = fn limit ->
-  "SELECT number FROM system.numbers_mt LIMIT #{limit}"
-end
-
-run_stream = fn statement, opts ->
-  f = fn conn -> conn |> Ch.stream(statement, [], opts) |> Stream.run() end
-  Ch.run(conn, f, timeout: :infinity)
-end
-
 Benchee.run(
   %{
-    "stream without decode" => fn statement ->
-      run_stream.(statement, _opts = [])
+    "RowBinary stream without decode" => fn limit ->
+      DBConnection.run(
+        conn,
+        fn conn ->
+          conn
+          |> Ch.stream(
+            "SELECT number FROM system.numbers_mt LIMIT {limit:UInt64} FORMAT RowBinary",
+            %{"limit" => limit}
+          )
+          |> Stream.run()
+        end,
+        timeout: :infinity
+      )
     end,
-    # TODO why is this faster?
-    "stream with manual decode" => fn statement ->
-      f = fn conn ->
-        conn
-        |> Ch.stream(statement, [], format: "RowBinary")
-        |> Stream.map(fn responses ->
-          Enum.each(responses, fn
-            {:data, _ref, data} -> Ch.RowBinary.decode_rows(data, [:u64])
-            {:status, _ref, 200} -> :ok
-            {:headers, _ref, _headers} -> :ok
-            {:done, _ref} -> :ok
+    "RowBinary stream with manual decode" => fn limit ->
+      DBConnection.run(
+        conn,
+        fn conn ->
+          conn
+          |> Ch.stream(
+            "SELECT number FROM system.numbers_mt LIMIT {limit:UInt64} FORMAT RowBinary",
+            %{"limit" => limit}
+          )
+          |> Stream.map(fn %Ch.Result{data: data} ->
+            data
+            |> IO.iodata_to_binary()
+            |> Ch.RowBinary.decode_rows([:u64])
           end)
-        end)
-        |> Stream.run()
-      end
-
-      Ch.run(conn, f, timeout: :infinity)
-    end,
-    "stream with decode" => fn statement ->
-      run_stream.(statement, types: [:u64])
+          |> Stream.run()
+        end,
+        timeout: :infinity
+      )
     end
   },
   inputs: %{
-    "500 rows" => statement.(500),
-    "500_000 rows" => statement.(500_000),
-    "500_000_000 rows" => statement.(500_000_000)
+    "500 rows" => 500,
+    "500_000 rows" => 500_000,
+    "500_000_000 rows" => 500_000_000
   }
 )