+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - add `:data` in `%Ch.Result{}` https://github.com/plausible/ch/pull/159
 - duplicate `Ch.Result.data` in `Ch.Result.rows` for backwards compatibility https://github.com/plausible/ch/pull/160
+- make `Ch.stream` emit `Ch.Result.t` instead of `Mint.Types.response`
 
 ## 0.2.4 (2024-01-29)